third_party/WebKit/LayoutTests/external/wpt/html/dom/dynamic-markup-insertion/document-write/write-active-document-expected.txt

This is a testharness.js-based test.
FAIL document.write only writes to active documentsassert_equals: expected (string) "IFRAME" but got (undefined) undefined
Harness: the test ran to completion.
